requirements.txt
Flask==3.0.0
numpy==1.26.4
pandas==2.2.1
tensorflow-cpu==2.15.0